Move files for x-pack move
